+++ modules/Examples/RecordShop-ExampleProject/RecordShop-service/src/test/java/org/company/recordshop/service/CustomerWithOrdersAndOrderLinesServiceTest.java
@@ -2,6 +2,7 @@ package org.company.recordshop.service;
 
 import static org.junit.Assert.assertEquals;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
@@ -138,7 +139,7 @@ public class CustomerWithOrdersAndOrderLinesServiceTest extends AbstractServiceT
         line.setLineNumber(i);
         FullProductDto product = new FullProductDto();
         product.setProductNumber(productNumber);
-        product.setPrice(15.0F);
+        product.setPrice(BigDecimal.valueOf(15.0));
         product.setOrderable(true);
         line.setProduct(product);
         return line;
